refactor: 공백 제거, 미사용 클래스 삭제
