+++ src/test/java/lotto/util/validator/InputValidatorTest.java
@@ -1,6 +1,5 @@
-package lotto.util;
+package lotto.util.validator;
 
-import lotto.util.validator.InputValidator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
